Create dataset downloading function
