+++ requirements.txt
@@ -4,4 +4,5 @@ git+https://github.com/facebookresearch/pytorch3d.git@v0.7.1
 numpy==1.23.5
 SimpleITK==2.2.0
 nibabel==4.0.2
-gdown~=4.5.3
+
+wandb==0.13.5
